Clean up reducer.js and rename
