+++ src/components/sets/cell-set-utils.js
@@ -45,131 +45,6 @@ function nodeToHeight(currNode, level = 0) {
 }
 
 /**
- * Set the ._state.level value for a node.
- * @param {object} currNode A node object.
- * @param {number} newLevel The level value.
- * @returns {object} The updated node.
- */
-/*
-function nodeSetLevel(currNode, newLevel) {
-  return {
-    ...currNode,
-    _state: {
-      ...currNode._state,
-      level: newLevel,
-    },
-  };
-}
-*/
-
-/**
- * Append a child to a parent node.
- * @param {object} currNode A node object.
- * @param {object} newChild The child node object.
- * @returns {object} The updated node.
- */
-/*
-function nodeAppendChild(currNode, newChild) {
-  const newChildWithLevel = nodeSetLevel(newChild, currNode._state.level + 1);
-  return {
-    ...currNode,
-    children: [...currNode.children, newChildWithLevel],
-  };
-}
-*/
-
-/**
- * Prepend a child to a parent node.
- * @param {object} currNode A node object.
- * @param {object} newChild The child node object.
- * @returns {object} The updated node.
- */
-/*
-function nodePrependChild(currNode, newChild) {
-  const newChildWithLevel = nodeSetLevel(newChild, currNode._state.level + 1);
-  return {
-    ...currNode,
-    children: [newChildWithLevel, ...currNode.children],
-  };
-}
-*/
-
-/**
- * Insert a child to a parent node.
- * @param {object} currNode A node object.
- * @param {*} newChild The child node object.
- * @param {*} insertIndex The index at which to insert the child.
- * @returns {object} The updated node.
- */
-/*
-function nodeInsertChild(currNode, newChild, insertIndex) {
-  const newChildWithLevel = nodeSetLevel(newChild, currNode._state.level + 1);
-  const newChildren = Array.from(currNode.children);
-  newChildren.splice(insertIndex, 0, newChildWithLevel);
-  return {
-    ...currNode,
-    children: newChildren,
-  };
-}
-*/
-
-
-/**
- * Find a node object that matches a predicate function.
- * @param {object} node A node object.
- * @param {function} predicate Returns true if a node matches a condition of interest.
- * @returns {object|null} A node object matching the predicate, or null if none is found.
- */
-/*
-function nodeFindNode(node, predicate) {
-  if (predicate(node)) {
-    return node;
-  }
-  if (!node.children) {
-    return null;
-  }
-  const foundNodes = node.children
-    .map(child => nodeFindNode(child, predicate))
-    .filter(Boolean);
-  if (foundNodes.length === 1) {
-    return foundNodes[0];
-  }
-  return null;
-}
-*/
-
-
-/**
- * Find a node matching a predicate function, relative to the whole tree.
- * @param {object} currTree A tree object.
- * @param {function} predicate Returns true if a node matches a condition of interest.
- * @returns {object|null} A matching node object, or null if none is found.
- */
-/*
-function treeFindNode(currTree, predicate) {
-  const foundNodes = currTree.tree
-    .map(levelZeroNode => nodeFindNode(levelZeroNode, predicate))
-    .filter(Boolean);
-  if (foundNodes.length === 1) {
-    return foundNodes[0];
-  }
-  return null;
-}
-*/
-
-/**
- * Find a node with a matching key, relative to the whole tree.
- * @param {object} currTree A tree object.
- * @param {string} targetKey The key for the node of interest.
- * @returns {object|null} A matching node object, or null if none is found.
- */
-/*
-function treeFindNodeByKey(currTree, targetKey) {
-  return treeFindNode(currTree, n => (n._state.key === targetKey));
-}
-*/
-
-/**
  * Find a node with a matching name path, relative to a particular node.
  * @param {object} node A node object.
  * @param {string[]} path The name path for the node of interest.
@@ -211,109 +86,6 @@ export function treeFindNodeByNamePath(currTree, targetNamePath) {
 }
 
 /**
- * Find a node with a matching name path, relative to the whole tree.
- * @param {object} currTree A tree object.
- * @param {string[]} targetNamePath The name path for the node of interest.
- * @returns {object|null} A matching node object, or null if none is found.
- */
-/*
-function treeFindNodeParentByNamePath(currTree, targetNamePath) {
-  const parentNamePath = [...targetNamePath];
-  parentNamePath.pop();
-  return treeFindNodeByNamePath(currTree, parentNamePath);
-}
-*/
-
-
-/**
- * Find the level zero node of a descendant node, using the descendant node's key.
- * @param {object} currTree A tree object.
- * @param {string} targetKey The key of the descendant node.
- * @returns {object|null} A matching level zero node object, or null if none is found.
- */
-/*
-function treeFindLevelZeroNodeByDescendantKey(currTree, targetKey) {
-  const predicate = n => (n._state.key === targetKey);
-  const foundNodes = currTree.tree
-    .map(levelZeroNode => (nodeFindNode(levelZeroNode, predicate)
-      ? levelZeroNode
-      : null))
-    .filter(Boolean);
-  if (foundNodes.length === 1) {
-    return foundNodes[0];
-  }
-  return null;
-}
-*/
-
-/**
- * Remove a node or any node descendants that match a predicate.
- * Recursive.
- * @param {object} node A node object.
- * @param {function} predicate Predicate function, returns true if the node should be removed.
- * @returns {object|null} The updated node, or null if the node matches the predicate.
- */
-/*
-function nodeRemove(node, predicate) {
-  if (predicate(node)) {
-    return null;
-  }
-  if (node.children) {
-    return {
-      ...node,
-      children: node.children
-        .map(child => nodeRemove(child, predicate))
-        .filter(Boolean),
-    };
-  }
-  return node;
-}
-*/
-
-/**
- * Remove any node matching a predicate, relative to the tree.
- * @param {object} currTree A tree object.
- * @param {string} targetKey The key of the node to remove.
- * @param {boolean} temporary Is this a temporary removal,
- * i.e. will the calling function ensure that the node will be added
- * back into the tree afterwards (for example, when dragging and dropping)?
- * By default, false.
- * @returns {object} The updated tree.
- */
-/*
-function treeNodeRemove(currTree, targetKey, temporary = false) {
-  const nodeToRemove = treeFindNodeByKey(currTree, targetKey);
-  const levelZeroNode = treeFindLevelZeroNodeByDescendantKey(currTree, targetKey);
-
-  const shouldClearCheckedLevel = (
-    (currTree._state.checkedLevel.levelZeroKey === nodeToRemove._state.key)
-    || (currTree._state.checkedLevel.levelZeroKey === levelZeroNode._state.key
-      && currTree._state.checkedLevel.levelIndex === nodeToRemove._state.level));
-
-  return {
-    ...currTree,
-    tree: currTree.tree
-      .map(node => nodeRemove(node, n => n._state.key === targetKey))
-      .filter(Boolean),
-    _state: {
-      ...currTree._state,
-      checkedKeys: (temporary
-        ? currTree._state.checkedKeys
-        : currTree._state.checkedKeys.filter(k => k !== targetKey)),
-      visibleKeys: (temporary
-        ? currTree._state.checkedKeys
-        : currTree._state.visibleKeys.filter(k => k !== targetKey)),
-      checkedLevel: (
-        !temporary && shouldClearCheckedLevel
-          ? { levelZeroKey: null, levelIndex: null }
-          : currTree._state.checkedLevel
-      ),
-    },
-  };
-}
-*/
-
-/**
  * Transform a node object using a transform function.
  * @param {object} node A node object.
  * @param {function} predicate Returns true if a node matches a condition of interest.
@@ -333,30 +105,6 @@ export function nodeTransform(node, predicate, transform) {
   }
   return node;
 }
-
-
-/**
- * Transform a node of interest, relative to the tree.
- * @param {object} currTree A tree object.
- * @param {string} targetKey The key for the node of interest.
- * @param {function} transform Takes the node of interest as input,
- * returns a transformed version of the node.
- * @returns {object} The updated tree.
- */
-/*
-function treeTransformNodeByKey(currTree, targetKey, transform) {
-  return {
-    ...currTree,
-    tree: currTree.tree
-      .map(node => nodeTransform(
-        node,
-        n => (n._state.key === targetKey),
-        transform,
-      )),
-  };
-}
-*/
-
 
 /**
  * Clear node state.
@@ -409,131 +157,6 @@ export function treeToComplement(currTree, checkedPaths, items) {
   const primaryUnion = treeToUnion(currTree, checkedPaths);
   return items.filter(el => !primaryUnion.includes(el));
 }
-
-/**
- * Respond to a node drag-and-drop interaction.
- * @param {object} currTree A tree object.
- * @param {string} dropKey The key of the node that the dragged node was dropped
- * onto or beside.
- * @param {string} dragKey The key of the node that was dragged.
- * @param {number} dropPosition The index at which to insert the dragged node,
- * if it was dropped beside the drop node.
- * @param {boolean} dropToGap Was the dragged node dropped onto the drop node?
- * @returns {object} The updated tree.
- */
-/* TODO: re-implement using node paths rather than keys.
-function treeOnDropNode(currTree, dropKey, dragKey, dropPosition, dropToGap) {
-  // Get drop node.
-  const dropNode = treeFindNodeByKey(currTree, dropKey);
-  const dropNodeLevel = dropNode._state.level;
-  const dropNodeIsLevelZero = dropNodeLevel === 0;
-  const dropNodeIsLevelZeroEmpty = (dropNodeIsLevelZero && (
-    !dropNode.children || dropNode.children.length === 0));
-  const dropNodeHeight = nodeToHeight(dropNode);
-  // Get drag node.
-  const dragNode = treeFindNodeByKey(currTree, dragKey);
-  const dragNodeHeight = nodeToHeight(dragNode);
-
-  // Only allow dragging if:
-  // - dropping between nodes, and both drag and drop node have same height, OR
-  // - dropping the dragNode into the dropNode,
-  //   where the dragNode has one less level than the dropNode.
-  // - dropping the dragNode into the dropNode,
-  //   where the dropNode is an _empty_ level zero node.
-  const isAllowed = (dropToGap && dropNodeHeight === dragNodeHeight)
-   || (!dropToGap && dropNodeHeight - 1 === dragNodeHeight) || dropNodeIsLevelZeroEmpty;
-
-  if (!isAllowed) {
-    return currTree;
-  }
-
-  let dropParentNode;
-  let dropParentKey;
-  let dropNodeCurrIndex;
-
-  if (!dropNodeIsLevelZero) {
-    dropParentNode = treeFindNodeParentByKey(currTree, dropKey);
-    dropParentKey = dropParentNode._state.key;
-    dropNodeCurrIndex = dropParentNode.children.findIndex(c => c._state.key === dropKey);
-  } else {
-    dropNodeCurrIndex = currTree.tree.findIndex(lzn => lzn._state.key === dropKey);
-  }
-
-  // Further, only allow dragging if the dragged node will have a unique
-  // name among its new siblings.
-  let hasSiblingNameConflict;
-  const dragNodeName = dragNode.name;
-  if (!dropNodeIsLevelZero) {
-    hasSiblingNameConflict = dropParentNode.children
-      .find(c => c.name === dragNodeName && c._state.key !== dragKey);
-  } else if (dropNodeIsLevelZero && !dropToGap) {
-    hasSiblingNameConflict = dropNode.children
-      .find(c => c.name === dragNodeName && c._state.key !== dragKey);
-  } else {
-    hasSiblingNameConflict = currTree.tree
-      .find(lzn => lzn.name === dragNodeName && lzn._state.key !== dragKey);
-  }
-
-  if (hasSiblingNameConflict) {
-    return currTree;
-  }
-
-  // Remove the dragged object from its current position.
-  const newTree = treeNodeRemove(currTree, dragKey, true);
-
-  // Update index values after temporarily removing the dragged node.
-  if (!dropNodeIsLevelZero) {
-    dropNodeCurrIndex = dropParentNode.children.findIndex(c => c._state.key === dropKey);
-  } else {
-    dropNodeCurrIndex = currTree.tree.findIndex(lzn => lzn._state.key === dropKey);
-  }
-
-  // Append the dragNode to dropNode's children if dropping _onto_ the dropNode.
-  if (!dropToGap) {
-    // Set dragNode as the last child of dropNode.
-    return treeTransformNodeByKey(newTree, dropKey, n => nodeAppendChild(n, dragNode));
-  }
-
-  // Prepend or insert the dragNode if dropping _between_ (above or below dropNode).
-  if (!dropNodeIsLevelZero) {
-    // The dropNode is at a level greater than zero,
-    // so it has a parent.
-    if (dropPosition === -1) {
-      // Set dragNode as first child of dropParentNode.
-      return treeTransformNodeByKey(
-        newTree,
-        dropParentKey,
-        n => nodePrependChild(n, dragNode),
-      );
-    }
-    // Set dragNode before or after dropNode.
-    const insertIndex = dropNodeCurrIndex + (dropPosition > dropNodeCurrIndex ? 1 : 0);
-    return treeTransformNodeByKey(
-      newTree,
-      dropParentKey,
-      n => nodeInsertChild(n, dragNode, insertIndex),
-    );
-  }
-  // We need to drop the dragNode to level zero,
-  // and level zero nodes do not have parents.
-  const newDragNode = nodeSetLevel(dragNode, 0);
-  if (dropPosition === -1) {
-    // Set dragNode as first level zero node of the tree.
-    return {
-      ...newTree,
-      tree: [newDragNode, ...newTree.tree],
-    };
-  }
-  // Set dragNode before or after dropNode in level zero.
-  const insertIndex = dropNodeCurrIndex + (dropPosition > dropNodeCurrIndex ? 1 : 0);
-  const newLevelZero = Array.from(newTree.tree);
-  newLevelZero.splice(insertIndex, 0, newDragNode);
-  return {
-    ...newTree,
-    tree: newLevelZero,
-  };
-}
-*/
 
 /**
  * Get an flattened array of descendants at a particular relative
